FIX: [9.2 RC] correct relate resources behaviour when upload_then_edit enabled [t22020] [t22104] [t21997]
git-svn-id: http://svn.resourcespace.com/svn/rs/releases/9.2_RC@14368 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/resource_functions.php
@@ -553,7 +553,7 @@ function save_resource_data($ref,$multi,$autosave_field="")
         }
       
    # Save related resource field if value for Related input field is autosaved, or if form has been submitted by user
-    if (($autosave_field=="") || ($autosave_field=="Related"))
+    if (($autosave_field=="" || $autosave_field=="Related") && isset($_POST["related"]))
         {
          # save related resources field
          sql_query("delete from resource_related where resource='$ref' or related='$ref'"); # remove existing related items
